chore: vector & tile 方法调整
